# [v2.9] Improve `go generate` performance with concurrency support