### Issue
The `go generate` command triggers `GenerateRegSyncFile()`, which generates the `regsync.yaml` file containing all images used by supported RKE, RKE2, and K3s Kubernetes versions in Rancher. To obtain the list of images for RKE2 and K3s, it downloads the `xx-images.txt` file from the GitHub releases of [RKE2](https://github.com/rancher/rke2/releases) and [K3s](https://github.com/k3s-io/k3s/releases). The downloads were performed sequentially, causing the main thread to wait for each image download to complete.

### Update
To improve performance and reduce wait times, concurrency support has been added to the `getImages()` function. This change allows for simultaneous downloads from multiple URLs, thereby avoiding blocking the main thread.

The code has been verified for race conditions using the `-race` go parameter to ensure thread safety.

**Time taken with sequential download:** 3m 23s

```
❯ time go generate ./...
WARN[0000] found k8s versions without a template        
WARN[0000] no aci template for k8sVersions [list of versions...]
finished generating data.json
INFO[0000] generating the regsync.yaml file             
INFO[0202] finished generating regsync.yaml             
go generate ./...  1.52s user 0.43s system 0% cpu 3:23.44 total
```

**Time taken with parallel download:** 5s

```
> time go generate ./...
WARN[0000] found k8s versions without a template        
WARN[0000] no aci template for k8sVersions [list of versions...]
finished generating data.json
INFO[0000] generating the regsync.yaml file             
INFO[0003] finished generating regsync.yaml             
go generate ./...  1.63s user 0.37s system 43% cpu 4.636 total
```