@chiukapoor 2024-10-15
> LGTM
> 
> Please backport / forwardport the fix after the PR is merged.

- dev-v2.9: https://github.com/rancher/kontainer-driver-metadata/pull/1519
- dev-v2.10: https://github.com/rancher/kontainer-driver-metadata/pull/1520